Refactor tests into separate files for clarity
