+++ polls/tests/test_detail_view.py
@@ -1,0 +1,24 @@
+import datetime
+
+from django.test import TestCase
+from django.utils import timezone
+from django.urls import reverse
+from polls.models import Question
+from http import HTTPStatus
+
+def create_question(question_text, days):
+    time = timezone.now() + datetime.timedelta(days = days)
+    return Question.objects.create(question_text = question_text, pub_date = time)
+
+class QuestionDetailViewTests(TestCase):
+    def test_future_question(self):
+        future_question = create_question(question_text="Future question." ,days=5)
+        url = reverse("polls:detail", args=(future_question.id,))
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, HTTPStatus.NOT_FOUND)
+    
+    def test_past_question(self):
+        past_question = create_question(question_text="Past question.", days=-5)
+        url = reverse("polls:detail", args=(past_question.id,))
+        response = self.client.get(url)
+        self.assertContains(response, past_question.question_text)
