+++ polls/tests/test_index_view.py
@@ -3,28 +3,11 @@ import datetime
 from django.test import TestCase
 from django.utils import timezone
 from django.urls import reverse
-from .models import Question
-from http import HTTPStatus
+from polls.models import Question
 
 def create_question(question_text, days):
     time = timezone.now() + datetime.timedelta(days = days)
     return Question.objects.create(question_text = question_text, pub_date = time)
-
-class QuestionModelTests(TestCase):
-    def test_was_published_recently_with_future_question(self):
-        future_time = timezone.now() + datetime.timedelta(days=30)
-        future_question = Question(pub_date = future_time)
-        self.assertIs(future_question.was_published_recently(), False)
-
-    def test_was_published_recently_with_old_question(self):
-        passed_time = timezone.now() - datetime.timedelta(days=1, seconds=1)
-        old_question = Question(pub_date = passed_time)
-        self.assertIs(old_question.was_published_recently(), False)
-    
-    def test_was_published_recently_with_recent_question(self):
-        recent_time = timezone.now() - datetime.timedelta(hours=23, minutes=59, seconds=59)
-        recent_question = Question(pub_date = recent_time)
-        self.assertIs(recent_question.was_published_recently(), True)
 
 class QuestionIndexViewTests(TestCase):
     def test_no_question(self):
@@ -54,16 +37,3 @@ class QuestionIndexViewTests(TestCase):
         past_question2 = create_question(question_text="Past question 2.", days=-5)
         response = self.client.get(reverse("polls:index"))
         self.assertQuerySetEqual(response.context["latest_question_list"], [past_question2, past_question1])
-
-class QuestionDetailViewTests(TestCase):
-    def test_future_question(self):
-        future_question = create_question(question_text="Future question." ,days=5)
-        url = reverse("polls:detail", args=(future_question.id,))
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, HTTPStatus.NOT_FOUND)
-    
-    def test_past_question(self):
-        past_question = create_question(question_text="Past question.", days=-5)
-        url = reverse("polls:detail", args=(past_question.id,))
-        response = self.client.get(url)
-        self.assertContains(response, past_question.question_text)
